+++ lib/generators/meta_tags/templates/config/initializers/meta_tags.rb
@@ -1,9 +1,14 @@
 # Use this setup block to configure all options available in MetaTags.
 MetaTags.configure do |config|
   # How many characters should the title meta tag have at most. Default is 70.
+  # Set to nil or 0 to remove limits.
   # config.title_limit = 70
 
+  # When true, site title will be truncated instead of title. Default is false.
+  # config.truncate_site_title_first = false
+
   # Maximum length of the page description. Default is 160.
+  # Set to nil or 0 to remove limits.
   # config.description_limit = 160
 
   # Maxumum length of the keywords meta tag. Default is 255.