+++ lib/meta_tags/configuration.rb
@@ -3,14 +3,22 @@ module MetaTags
   class Configuration
     # How many characters to truncate title to.
     attr_accessor :title_limit
+
+    # Truncate site_title instead of title.
+    attr_accessor :truncate_site_title_first
+
     # How many characters to truncate description to.
     attr_accessor :description_limit
+
     # How many characters to truncate keywords to.
     attr_accessor :keywords_limit
+
     # Keywords separator - a string to join keywords with.
     attr_accessor :keywords_separator
+
     # Should keywords forced into lowercase?
     attr_accessor :keywords_lowercase
+
     # Custom meta tags that should use `property` attribute instead of `name`
     # - an array of strings or symbols representing their names or name-prefixes.
     attr_reader :property_tags
@@ -58,6 +66,7 @@ module MetaTags
 
     def reset_defaults!
       @title_limit = 70
+      @truncate_site_title_first = false
       @description_limit = 160
       @keywords_limit = 255
       @keywords_separator = ', '